cleanup: move pkg/tool/cache to pkg/util/cache
