+++ pkg/koordlet/executor/resource_update_executor.go
@@ -22,7 +22,7 @@ import (
 
 	"k8s.io/klog/v2"
 
-	"github.com/koordinator-sh/koordinator/pkg/tools/cache"
+	"github.com/koordinator-sh/koordinator/pkg/util/cache"
 )
 
 var _ CacheExecutor = &ResourceUpdateExecutor{}
