+++ pkg/koordlet/resmanager/resmanager.go
@@ -20,8 +20,6 @@ import (
 	"context"
 	"fmt"
 
-	runtime "github.com/koordinator-sh/koordinator/pkg/koordlet/util/runtime"
-
 	corev1 "k8s.io/api/core/v1"
 	policyv1beta1 "k8s.io/api/policy/v1beta1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -44,8 +42,9 @@ import (
 	"github.com/koordinator-sh/koordinator/pkg/koordlet/resmanager/configextensions"
 	"github.com/koordinator-sh/koordinator/pkg/koordlet/resmanager/plugins"
 	"github.com/koordinator-sh/koordinator/pkg/koordlet/statesinformer"
-	expireCache "github.com/koordinator-sh/koordinator/pkg/tools/cache"
+	"github.com/koordinator-sh/koordinator/pkg/koordlet/util/runtime"
 	"github.com/koordinator-sh/koordinator/pkg/util"
+	expireCache "github.com/koordinator-sh/koordinator/pkg/util/cache"
 )
 
 const (
