+++ pkg/koordlet/util/resourceexecutor/executor.go
@@ -24,7 +24,7 @@ import (
 
 	"k8s.io/klog/v2"
 
-	"github.com/koordinator-sh/koordinator/pkg/tools/cache"
+	"github.com/koordinator-sh/koordinator/pkg/util/cache"
 )
 
 var _ ResourceUpdateExecutor = &ResourceUpdateExecutorImpl{}
